Split into modules and publicize the JIF structs
All the parser/struct things have been moved to jpeg, and the
`parse_with` helpers to `helpers`. Additionally, all the jpeg structs
and their fields have been made public, since that's going to be part of
the point of this once it's done: a library for poking at jpeg-guts in
addition to an application for doing the same.

This change has the added benefit of getting rid of all the warnings
about the many unused fields