Implement ECS restart interval parsing
The parser can parse ECS segments interspersed with RST markers,but does
no validation, simply looking for RST markers breaking up the ECS data.
It does not care if there is or is not a DRI, and it does not care about
the size of the ECS blocks nor their consistency.