Parse contents of DQT segments fully
We use the try_map directive to limit the number of bytes to read,
combined with the until_invalid helper to try and read tables until we
run out of that limited number of bytes.

There's maybe? probably? a way to figure out how many tables there are
based on other pieces in the file, but it also seems like it's valid to
use multiple DQT segments to define each of your tables, or to split
them up arbitrarily, in which case... this is all we've got. Besides,
libjpeg seems to do it this way too (if more imperatively)