Implement huffman and MCU/component decoding
Move the hand-testing, partially implemented decoding & huffman
decoding parts that were in main into their own modules. Additionally,
draw the rest of the owl.

The owl includes: huffman decoding, jpeg number decoding (for the SSSS
stuff), MCU disentangling, IDCT that might work, and some internal
infrastructure for reading bits and placing samples on a grid

Output is currently idosyncratic: with gridlines and some sample
clipping on a test image, but it does actually correspond to the source
image!